Fixing for 64 bit windows compilation

+++ src/snaplib/dbl4/snap_dbl4_interface.c
@@ -143,7 +143,7 @@ StatusType utlSeekBlobDB( void *blob, long lngOffset, int whence )
 
 StatusType utlTellBlobDB( void *blob, long *lngOffset )
 {
-    (*lngOffset) = ftell( (FILE *) blob );
+    (*lngOffset) = (long) ftell( (FILE *) blob );
     return STS_OK;
 }
